chore(mcp): extract aether-rpc to drop the native aether-capabilities edge
