+++ crates/aether-rpc/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "aether-rpc"
+version.workspace = true
+edition.workspace = true
+
+# ADR-0102: the RPC wire vocabulary + the `Call` client primitive,
+# extracted out of `aether-capabilities` so they live in a crate with no
+# path to `aether-substrate` / wasmtime / wgpu. The MCP coordinator
+# (`aether-mcp`) depends on this crate alone for its production surface;
+# `aether-capabilities` re-exports these types at their original paths
+# (`aether_capabilities::rpc::*`, `aether_capabilities::trace_walk::*`)
+# and keeps the substrate-bound `RpcServerCapability` next to them. The
+# crate edge is the mechanism that holds the coordinator's "no native
+# deps" invariant under cargo's workspace-wide feature unification — a
+# feature flag cannot (ADR-0102 §Context).
+
+[dependencies]
+# Typed wire ids (`EngineId`, `KindId`, `MailboxId`, `ThreadId`,
+# `MailId`) the wire envelopes + the trace walk are keyed on. `no_std` +
+# `alloc`, no native deps.
+aether-data = { path = "../aether-data" }
+# Trace vocabulary the guided walk reconstructs (`TraceRingEntry`,
+# `MailNodeWire`, `DescribeTreeResult`, `TraceEvent`, `Nanos`).
+aether-kinds = { path = "../aether-kinds" }
+# Length-prefix postcard stream framing the `Call` client reads/writes
+# (ADR-0072). Reached only by the native-only client; harmless in the
+# wasm-header build (`aether-capabilities` already pulls it there).
+aether-codec = { path = "../aether-codec" }
+serde = { workspace = true, default-features = false, features = ["derive", "alloc"] }
+
+[lints]
+workspace = true
